Create graph control panel (#403)
* Create graph control panel component

* Move buttons to control panel

* Fix reset and move origin button

* Fix indentation
