+++ src/shared/components/Graph/GraphComponent.less
@@ -3,24 +3,7 @@
 .graph-component {
   height: 100%;
   width: 100%;
-  & > .top {
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    margin-top: 1em;
-    button {
-      margin-right: 4px;
-    }
-    .alert {
-      margin: 1em;
-    }
-    .controls {
-      justify-content: space-between;
-      display: flex;
-      width: 100%;
-    }
-  }
+  
   & > .graph {
     width: 100%;
     height: 100%;
