+++ src/shared/components/Graph/index.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import * as cytoscape from 'cytoscape';
-import { Alert, Button } from 'antd';
 import * as cola from 'cytoscape-cola';
 
 import style from './style';
@@ -26,26 +25,17 @@ const Graph: React.FunctionComponent<{
   onNodeClick?(id: string, data: ElementNodeData): void;
   onNodeClickAndHold?(id: string, data: ElementNodeData): void;
   onNodeHover?(id: string, idata: ElementNodeData): void;
-  onReset?(): void;
-  onCollapse?(): void;
-  onLayoutChange?(type: string): void;
   layout?: string;
-  collapsed?: boolean;
-  loading: boolean;
+  centered?: boolean;
 }> = ({
   elements,
   onNodeClick,
   onNodeClickAndHold,
   onNodeHover,
-  onReset,
-  collapsed,
-  onCollapse,
-  onLayoutChange,
   layout = DEFAULT_LAYOUT,
-  loading,
+  centered,
 }) => {
   const container = React.useRef<HTMLDivElement>(null);
-  const [showAlert, setShowAlert] = React.useState(true);
   const [cursorPointer, setCursorPointer] = React.useState<string | null>(null);
   const layoutInstance = React.useRef<cytoscape.Layouts>();
   const graph = React.useRef<cytoscape.Core>();
@@ -67,10 +57,6 @@ const Graph: React.FunctionComponent<{
         })
         .run();
     }
-  };
-
-  const handleLayoutClick = (type: string) => () => {
-    onLayoutChange && onLayoutChange(type);
   };
 
   const onRecenter = () => {
@@ -193,6 +179,10 @@ const Graph: React.FunctionComponent<{
   }, [JSON.stringify(elements)]);
 
   React.useEffect(() => {
+    onRecenter();    
+  }, [centered]);
+
+  React.useEffect(() => {
     cytoscape.use(cola);
     graph.current = cytoscape({
       elements,
@@ -217,55 +207,6 @@ const Graph: React.FunctionComponent<{
         style={cursorPointer ? { cursor: cursorPointer } : {}}
       ></div>
       <GraphLegend />
-      <div className="top">
-        <div className="controls">
-          <div>
-            {Object.keys(LAYOUTS).map(layoutKey => {
-              return (
-                <Button
-                  key={layoutKey}
-                  size="small"
-                  type={layoutKey === layout ? 'primary' : 'default'}
-                  onClick={handleLayoutClick(layoutKey)}
-                >
-                  {LAYOUTS[layoutKey].label}
-                </Button>
-              );
-            })}
-          </div>
-          <div>
-            <Button
-              type={collapsed ? 'primary' : 'default'}
-              size="small"
-              onClick={onCollapse}
-            >
-              {collapsed ? 'Expand Paths' : 'Collapse Paths'}
-            </Button>
-            <Button size="small" onClick={onRecenter}>
-              Origin
-            </Button>
-            <Button size="small" onClick={onReset}>
-              Reset
-            </Button>
-          </div>
-        </div>
-        {showAlert ? (
-          <Alert
-            style={{ margin: '7px 5px 0 0' }}
-            message="Click and hold to visit a resource"
-            type="info"
-            closable
-            afterClose={() => setShowAlert(false)}
-          />
-        ) : null}
-        {loading && (
-          <Alert
-            style={{ margin: '7px 5px 0 0' }}
-            message="Loading..."
-            type="info"
-          />
-        )}
-      </div>
     </div>
   );
 };
